updating version

+++ inst/include/CGAL/CORE/extLong_impl.h
@@ -15,17 +15,17 @@
  *      Level i representation of a number n is just i iterations
  *      of log_2 applied to n.
  *
- * Written by 
+ * Written by
  *       Chee Yap <yap@cs.nyu.edu>
  *       Chen Li <chenli@cs.nyu.edu>
  *       Zilin Du <zilin@cs.nyu.edu>
- *       Sylvain Pion <pion@cs.nyu.edu> 
+ *       Sylvain Pion <pion@cs.nyu.edu>
  *
  * WWW URL: http://cs.nyu.edu/exact/
  * Email: exact@cs.nyu.edu
  *
- * $URL: https://github.com/CGAL/cgal/blob/releases/CGAL-5.0/CGAL_Core/include/CGAL/CORE/extLong_impl.h $
- * $Id: extLong_impl.h 26fb266 2019-10-19T16:28:33+02:00 Sébastien Loriot
+ * $URL: https://github.com/CGAL/cgal/blob/v5.2.1/CGAL_Core/include/CGAL/CORE/extLong_impl.h $
+ * $Id: extLong_impl.h 627a7b2 2020-04-27T17:08:35+02:00 Maxime Gimeno
  * SPDX-License-Identifier: LGPL-3.0-or-later
  ***************************************************************************/
 
@@ -37,7 +37,7 @@
 
 #include <CGAL/CORE/extLong.h>
 
-namespace CORE { 
+namespace CORE {
 
 CGAL_INLINE_FUNCTION
 const extLong& extLong::getNaNLong() {
@@ -125,9 +125,9 @@ extLong& extLong::operator*= (const extLong& y) {
     if (std::fabs(d - p) <= std::fabs(d) * relEps) {
       val = p;
       flag = 0;
-    } else if (d > EXTLONG_MAX) {
+    } else if (d > static_cast<double>(EXTLONG_MAX)) {
       *this = CORE_posInfty;
-    } else if (d < EXTLONG_MIN) {
+    } else if (d < static_cast<double>(EXTLONG_MIN)) {
       *this = CORE_negInfty;
     } else {
 #ifdef CORE_DEBUG
